Move provincial admin pages
